# Instead of the clock algorithm, always evict the largest layer.