This avoids nasty behavior where one large layer would hog all the
memory, causing many many other small layers to be evicted, which leads
to write amplification and bad performance. That was quite visible in
a plain pgbench workload.