@hlinnaka 2021-12-03
This is no longer relevant after #864, closing